refactor: pull HTML page rendering to a module
